update readme and adding cart
