+++ README.md
@@ -1,2 +1,20 @@
 [Portugues](/doc/pt-br.md)
 [Español](/doc/es.md)
+
+# Challenge M3
+
+## NOTA
+
+Si se ejecuta en MacOS: 
+* cambiar el puerto del json-server  en el `package.json` ya que el 5000 lo usa el sistema
+* actualizar la url en la linea 40 del archivo `index.js`.
+
+## Instrucciones
+Luego de clonarlo:
+
+1. Para instalarlo `npm install`
+2. Para ejecutarlo  `npm start`
+
+### Contacto
+
+email: emrahedi@live.com
